requirements.txt
Package             Version
------------------- -----------
alembic             1.8.1
anyio               3.6.1
asgiref             3.5.2
autopep8            1.6.0
bcrypt              3.2.2
certifi             2022.5.18.1
cffi                1.15.1
charset-normalizer  2.0.12
click               8.1.3
cryptography        37.0.4
dnspython           2.2.1
ecdsa               0.18.0
email-validator     1.2.1
fastapi             0.78.0
greenlet            1.1.2
h11                 0.13.0
httpcore            0.16.3
httptools           0.4.0
httpx               0.23.3
idna                3.3
importlib-metadata  4.12.0
importlib-resources 5.9.0
itsdangerous        2.1.2
Jinja2              3.1.2
Mako                1.2.1
MarkupSafe          2.1.1
orjson              3.7.2
passlib             1.7.4
pip                 23.0.1
psycopg2-binary     2.9.3
pyasn1              0.4.8
pycodestyle         2.8.0
pycparser           2.21
pydantic            1.9.1
python-dotenv       0.20.0
python-jose         3.3.0
python-multipart    0.0.5
PyYAML              6.0
requests            2.28.0
rfc3986             1.5.0
rsa                 4.9
setuptools          63.2.0
six                 1.16.0
sniffio             1.2.0
SQLAlchemy          2.0.7
starlette           0.19.1
toml                0.10.2
tomli               2.0.1
typing_extensions   4.2.0
ujson               5.3.0
urllib3             1.26.9
uvicorn             0.17.6
uvloop              0.16.0
watchfiles          0.18.1
watchgod            0.8.2
websockets          10.3
zipp                3.8.1
